pyproject.toml
[tool.poetry]
name = "djangochat"
version = "0.1.0"
description = ""
authors = ["EgorLenko <yehorlenko@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.5"
psycopg2 = "^2.9.5"
djongo = "^1.3.6"
channels = {extras = ["daphne"], version = "^4.0.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
